Backport #13430 to 8.0 Replace generated cleaner code with concrete implementation
Backport #13430 to 8.0 branch. Original message:

* Replace generated cleaner code with concrete implementation

No need to to perform codegen to create JDK specific compatible classes to
call the ByteBuffer cleaner, as we are dropping support for Java 8.

* Remove dead/redundant java 8 code from LogstashJavaCompat

Removes setupByteBuffer, specific test for Java9 and above and updates memory monitor
unit test to only test for values relevant to Java>1.8

* Renamed version utility class to reflect use after removal of methods

Also fixed some dead imports
