+++ logstash-core/src/main/java/org/logstash/ackedqueue/io/ByteBufferCleanerImpl.java
@@ -1,0 +1,26 @@
+package org.logstash.ackedqueue.io;
+
+import java.nio.MappedByteBuffer;
+
+import sun.misc.Unsafe;
+import java.lang.reflect.Field;
+
+public class ByteBufferCleanerImpl implements ByteBufferCleaner {
+
+    private final Unsafe unsafe;
+
+    public ByteBufferCleanerImpl() {
+        try {
+            Field unsafeField = Unsafe.class.getDeclaredField("theUnsafe");
+            unsafeField.setAccessible(true);
+            unsafe = (Unsafe) unsafeField.get(null);
+        }catch (Exception e){
+            throw new IllegalStateException(e);
+        }
+    }
+
+    @Override
+    public void clean(MappedByteBuffer buffer) {
+        unsafe.invokeCleaner(buffer);
+    }
+}
