+++ logstash-core/src/main/java/org/logstash/ackedqueue/io/MmapPageIOV2.java
@@ -32,7 +32,6 @@ import java.util.List;
 import java.util.zip.CRC32;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.logstash.LogstashJavaCompat;
 import org.logstash.ackedqueue.SequencedList;
 
 /**
@@ -54,8 +53,7 @@ public final class MmapPageIOV2 implements PageIO {
     /**
      * Cleaner function for forcing unmapping of backing {@link MmapPageIOV2#buffer}.
      */
-    private static final ByteBufferCleaner BUFFER_CLEANER =
-        LogstashJavaCompat.setupBytebufferCleaner();
+    private static final ByteBufferCleaner BUFFER_CLEANER = new ByteBufferCleanerImpl();
 
     private final File file;
